Move CatalogTree from views to components/connection
