+++ airbyte-webapp/src/components/connection/CatalogTree/Arrow.tsx
