+++ airbyte-webapp/src/components/connection/CatalogTree/BulkHeader.module.scss
