+++ airbyte-webapp/src/components/connection/CatalogTree/CatalogSection.module.scss
@@ -1,4 +1,4 @@
-@use "../../../scss/colors";
+@use "scss/colors";
 
 .streamFieldTableContainer {
   margin-left: 85px;
