+++ airbyte-webapp/src/components/connection/CatalogTree/CatalogTreeBody.module.scss
