+++ airbyte-webapp/src/components/connection/CatalogTree/CatalogTreeHeader.module.scss
