+++ airbyte-webapp/src/components/connection/CatalogTree/CatalogTreeHeader.tsx
