+++ airbyte-webapp/src/components/connection/CatalogTree/CatalogTreeSubheader.module.scss
