+++ airbyte-webapp/src/components/connection/CatalogTree/PathPopout.module.scss
