+++ airbyte-webapp/src/components/connection/CatalogTree/PathPopout.tsx
