+++ airbyte-webapp/src/components/connection/CatalogTree/PathPopoutButton.module.scss
@@ -1,5 +1,5 @@
-@use "../../../../scss/colors";
-@use "../../../../scss/variables";
+@use "scss/colors";
+@use "scss/variables";
 
 .arrow {
   color: colors.$grey-300;
