+++ airbyte-webapp/src/components/connection/CatalogTree/PathPopoutButton.tsx
