+++ airbyte-webapp/src/components/connection/CatalogTree/StreamHeader.module.scss
@@ -1,5 +1,5 @@
-@use "../../../scss/colors";
-@use "../../../scss/variables";
+@use "scss/colors";
+@use "scss/variables";
 @forward "./CatalogTreeBody.module.scss";
 
 .icon {
