+++ airbyte-webapp/src/components/connection/CatalogTree/SyncSettingsDropdown.tsx
