+++ airbyte-webapp/src/components/connection/CatalogTree/styles.tsx
